Change tests to adjust for delay from server
Some of the tests were changed so that they better handle
the fact that the mock server takes a second to respond with
the data requested on page load.

The mock API request handler was modified to wait a few
seconds before sending the response. This better simulates a
real server taking a few seconds to respond with assignment data.