Define components to represent columns checkboxes
Some new components were defined to display checkboxes that
allow the user to select which columns are displayed. A new
array of objects representing the labels for each of these
checkboxes was also created. When the app is initially mounted,
it produces the checkboxes dynamically based on the objects in
this array.