Refactor Columns modal and creat Filters modal
The Columns modal was refactored so that the same Modal
component could be used to display different content when the
user clicks the "Filters" button. This reduces the amount of
redundant code that will be written later.

A new Filters button was added to the page. When a user clicks
it, a Filters modal appears.